Simpler analyze entry-point for tests and main
for #70
